feat: split ether lib
