+++ flows/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rfq-marketplace"
+name = "rfq_marketplace"
 version = "0.1.0"
 edition = "2021"
 
@@ -12,7 +12,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 log = "0.4"
-http_req_wasi = { version = "0.10", features = ["wasmedge_ssl"] }
+http_req_wasi = { version = "0.11.1"}
 flowsnet-platform-sdk = "0.1"
 tokio_wasi = { version = "1.25.1", features = ["macros", "rt", "sync"] }
 webhook-flows = "0.4.4"
